+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v0.3.3]
 - encode error as header  [#10](https://github.com/xmidt-org/argus/pull/10)
 - add ability to disable pullInterval for chrysom  [#10](https://github.com/xmidt-org/argus/pull/10)
 - remove authHeader from logging [#10](https://github.com/xmidt-org/argus/pull/10)
@@ -38,7 +40,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [v0.1.0] Tue May 07 2020 Jack Murdock - 0.1.0
 - initial creation
 
-[Unreleased]: https://github.com/xmidt-org/argus/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/xmidt-org/argus/compare/v0.3.3...HEAD
+[v0.3.3]: https://github.com/xmidt-org/argus/compare/v0.3.2...v0.3.3
 [v0.3.2]: https://github.com/xmidt-org/argus/compare/v0.3.1...v0.3.2
 [v0.3.1]: https://github.com/xmidt-org/argus/compare/v0.3.0...v0.3.1
 [v0.3.0]: https://github.com/xmidt-org/argus/compare/v0.2.1...v0.3.0